Disable autodoc until typehints fixed

+++ docs/source/conf.py
@@ -28,34 +28,35 @@ sys.path.insert(0, os.path.join(__location__, "../.."))
 # setup.py install" in the RTD Advanced Settings.
 # Additionally it helps us to avoid running apidoc manually
 
-try:  # for Sphinx >= 1.7
-    from sphinx.ext import apidoc
-except ImportError:
-    from sphinx import apidoc
+if False:
+    try:  # for Sphinx >= 1.7
+        from sphinx.ext import apidoc
+    except ImportError:
+        from sphinx import apidoc
 
-output_dir = os.path.join(__location__, "api/auto")
-module_dir = os.path.join(__location__, "../../pycollo")
-try:
-    shutil.rmtree(output_dir)
-except FileNotFoundError:
-    pass
+    output_dir = os.path.join(__location__, "api/auto")
+    module_dir = os.path.join(__location__, "../../pycollo")
+    try:
+        shutil.rmtree(output_dir)
+    except FileNotFoundError:
+        pass
 
-try:
-    import sphinx
+    try:
+        import sphinx
 
-    cmd_line_template = (
-        "sphinx-apidoc --implicit-namespaces -f -o {outputdir} {moduledir}"
-    )
-    cmd_line = cmd_line_template.format(outputdir=output_dir, moduledir=module_dir)
+        cmd_line_template = (
+            "sphinx-apidoc --implicit-namespaces -f -o {outputdir} {moduledir}"
+        )
+        cmd_line = cmd_line_template.format(outputdir=output_dir, moduledir=module_dir)
 
-    args = cmd_line.split(" ")
-    if tuple(sphinx.__version__.split(".")) >= ("1", "7"):
-        # This is a rudimentary parse_version to avoid external dependencies
-        args = args[1:]
+        args = cmd_line.split(" ")
+        if tuple(sphinx.__version__.split(".")) >= ("1", "7"):
+            # This is a rudimentary parse_version to avoid external dependencies
+            args = args[1:]
 
-    apidoc.main(args)
-except Exception as e:
-    print("Running `sphinx-apidoc` failed!\n{}".format(e))
+        apidoc.main(args)
+    except Exception as e:
+        print("Running `sphinx-apidoc` failed!\n{}".format(e))
 
 
 # -- Project information -----------------------------------------------------
